migrations/requirements.txt
psycopg2
python-dotenv
jinja2
python-multipart
requests